vfs0090: Add support for in-chip database matching
If a fingerprint has been already enrolled in chip (using windows, or
external tools), we can safely use it.

When enrolling, we just associate it to the fprint print, in the
verification and identification cases we check if the print is device-saved
and in such case we compare is if its ID is the same of the one associated
to the internal print.

If such tests fails, we just fall-back to the image verification, if
possible.

This opens the gates to 0097 chip as well, that at this point is almost
supported, given that this code works even in case the device is not capable
of doing full image capture